Move capture_stdout into own module
